Split functions and cleanup extra lines
